// src/models/public/alumnos.model.ts
export interface IPublicAlumno{
    periodo: string
    zonal: string
    campus: string
    semestre: string
    batch_uid: string
    alumno:string
    nrc: string
    matcur:string
    listacruzada:string
    course_id:string
    curso:string
    fecha_acad: string
    condicion_acad:string
    ultimo_acceso: string
    pago:string
    deuda: string
    apto:string
    matriculable:string
    usuarioenrolado: string
    usuariovisiblecurso:string
    usuariovisiblebb:string
}